Ensure unsigned enum initializers cast to the signed underlying type
An in-range unsigned initializer on an enum with a signed underlying type (e.g.
`1U << 22` on the default `int`) was emitted without a cast, producing a
`uint` assigned to an `int` member which fails to compile (CS0266). Emit an
explicit cast to the underlying type in that case, matching the existing
out-of-range `unchecked((int)...)` behavior.

+++ tests/ClangSharp.PInvokeGenerator.UnitTests/Baseline/Baselines/EnumDeclaration/WithAnonymousEnumTest.CSharp.Preview.Unix.cs
@@ -4,7 +4,7 @@ namespace ClangSharp.Test
 {
     public enum MyEnum2
     {
-        MyEnum2_Value1 = MyEnum1_Value1,
+        MyEnum2_Value1 = ((int)(MyEnum1_Value1)),
     }
 
     public static partial class Methods
